webD_Backend_math-a-hack/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
pytz==2020.1
sqlparse==0.3.1
jsonfield==3.1.0
django-tinymce==3.0.2
whitenoise==5.1.0
mysqlclient==2.0.1
docutils==0.16